+++ website/src/components/Hero.vue
@@ -1,18 +1,17 @@
 <template>
     <section class="hero">
-        <!-- <v-layout class="overflow-visible" style="height: 56px; width: 500px"> -->
         <div style="margin-top: 8px;">
             <v-btn-toggle v-model="switchTheme" color="primary" variant="outlined" style="min-width: 200px;">
-                <v-btn><span class="hidden-sm-and-down">Dark</span></v-btn>
-                <v-btn><span class="hidden-sm-and-down">Light</span></v-btn>
+                <v-btn size="x-small" density="default" width="64">{{ $t('app.theme.dark') }}</v-btn>
+                <v-btn size="x-small" density="default" width="64">{{ $t('app.theme.light') }}</v-btn>
             </v-btn-toggle>
             <span style="margin: 24px;" />
             <v-btn-toggle v-model="switchLocale" color="primary" variant="outlined" style="min-width: 200px;">
-                <v-btn v-for="locale in $i18n.availableLocales" :key="`locale-${locale}`" :value="locale">
-                    <span class="hidden-sm-and-down">{{ locale }}</span>
+                <v-btn v-for="locale in $i18n.availableLocales" :key="`locale-${locale}`" :value="locale" size="x-small"
+                    density="default" width="48">
+                    {{ locale }}
                 </v-btn>
             </v-btn-toggle>
-            <!-- </v-layout> -->
         </div>
         <div class="hero-container">
             <div class="preview">
@@ -30,7 +29,7 @@
                         {{ $t('hero.caption3') }}
                     </div>
                     <br />
-                    <v-btn rounded="xl" size="x-large" block @click="scrollToDownload">
+                    <v-btn rounded="xl" size="x-large" color="#3b87af" block @click="scrollToDownload">
                         {{ $t('hero.getTammy') }}
                     </v-btn>
                 </div>
@@ -42,13 +41,12 @@
 
 <script lang="ts">
 import { useTheme } from 'vuetify'
-import { computed, watch, defineProps } from 'vue'
 import { useI18n } from 'vue-i18n'
-import { getCurrentInstance } from 'vue'
 import previewImageDark from '../assets/app-preview-dark.png'
 import previewImageLight from '../assets/app-preview-light.png'
 var theme = null;
 var i18n = null;
+
 export default {
     props: {
         goToDownloadSection: {
@@ -86,7 +84,10 @@ export default {
             console.log("locale " + newValue)
             if (!!i18n && !!newValue) i18n.locale.value = newValue
         },
-    }
+    },
+    // computed: {
+    //     getColorCaption3: () => (!!theme && !!theme.global.current.value.dark) ? '#ff0000' : '#00ff00',
+    // },
 }
 </script>
 
@@ -121,7 +122,8 @@ export default {
     }
 
     .caption-3 {
-        color: #3c3b3b;
+        /* color: v-bind('getColorCaption3'); */
+        color: #8c8c8c;
         font-size: 1.6em;
         font-weight: 300;
         width: 300px;